requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
python-dotenv==1.0.0
google-api-python-client==2.108.0
youtube-transcript-api==1.2.3
alembic==1.12.1
websockets==12.0
fabric>=3.0.0
feedparser>=6.0.11
apscheduler>=3.10.4